# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  viewer: 'viewer',
  editor: 'editor',
} as const;

export interface AuthUser {
  entraOid: string;
  email: string;
  displayName: string;
  role: AuthUserRole;
}

export type PackingOrderTeam = typeof PackingOrderTeam[keyof typeof PackingOrderTeam];


export const PackingOrderTeam = {
  PARTS: 'PARTS',
  SYSTEM: 'SYSTEM',
} as const;

export interface PackingOrder {
  id: string;
  team: PackingOrderTeam;
  salesOrder: string;
  isRetailOrder: boolean;
  customer: string;
  productDescription: string;
  machineModel: string;
  confirmedShipDate: string;
  requestedShipDate: string;
  customerPO: string;
  modeOfDelivery: string;
  incoterms: string;
  namedPlace: string;
  carrierForwarder: string;
  carrierForwarderDescription: string;
  customerCarrierAccount: string;
  shippingInstructions: string;
  freightInformation: string;
  documentStatus: string;
  salesperson: string;
  country: string;
  deliveryCity: string;
  deliveryState: string;
  customerReference: string;
  createdDate: string;
  inPackingAt: string;
  doNotProcess: boolean;
  doNotProcessKnown: boolean;
  /** @nullable */
  packStartedAt: string | null;
}

export type PackingOrdersResponseSource = typeof PackingOrdersResponseSource[keyof typeof PackingOrdersResponseSource];


export const PackingOrdersResponseSource = {
  vw_packing_control_board: 'vw_packing_control_board',
  salesorderheaderv3staging: 'salesorderheaderv3staging',
} as const;

export interface PackingOrdersResponse {
  source: PackingOrdersResponseSource;
  fetchedAt: string;
  /** @minimum 0 */
  skippedRows: number;
  orders: PackingOrder[];
}

export interface PackStatus {
  orderId: string;
  /** @nullable */
  packStartedAt: string | null;
}

export interface PackStatusesResponse {
  statuses: PackStatus[];
}

export interface PackStatusInput {
  /** @nullable */
  packStartedAt: string | null;
}

export type LoginParams = {
returnTo?: string;
};

export type AuthCallbackParams = {
code?: string;
state?: string;
error?: string;
};