# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  entraOid: string;
  email: string;
  displayName: string;
  role: AuthUserRole;
}