# lib/api-zod/src/generated/types/packStatusInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PackStatusInput {
  /** @nullable */
  packStartedAt: Date | null;
}