# lib/api-zod/src/generated/types/packStatusesResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PackStatus } from './packStatus';

export interface PackStatusesResponse {
  statuses: PackStatus[];
}