# lib/api-zod/src/generated/types/packingOrder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PackingOrderTeam } from './packingOrderTeam';

export interface PackingOrder {
  id: string;
  team: PackingOrderTeam;
  salesOrder: string;
  isRetailOrder: boolean;
  customer: string;
  productDescription: string;
  machineModel: string;
  confirmedShipDate: string;
  requestedShipDate: string;
  customerPO: string;
  modeOfDelivery: string;
  incoterms: string;
  namedPlace: string;
  carrierForwarder: string;
  carrierForwarderDescription: string;
  customerCarrierAccount: string;
  shippingInstructions: string;
  freightInformation: string;
  documentStatus: string;
  salesperson: string;
  country: string;
  deliveryCity: string;
  deliveryState: string;
  customerReference: string;
  createdDate: string;
  inPackingAt: Date;
  doNotProcess: boolean;
  doNotProcessKnown: boolean;
  /** @nullable */
  packStartedAt: Date | null;
}